% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alien.R
\docType{data}
\name{alien_theme_colors}
\alias{alien_theme_colors}
\title{ALIEN Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
alien_theme_colors
}
\description{
ALIEN Theme Color Palette
}
\examples{
alien_theme_colors

}
\keyword{datasets}
