% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wes_anderson.R
\docType{data}
\name{asteroid_city_theme_colors}
\alias{asteroid_city_theme_colors}
\title{Asteroid City Inspired Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
asteroid_city_theme_colors
}
\description{
Asteroid City Inspired Theme Color Palette
}
\examples{
asteroid_city_theme_colors

}
\concept{wes_anderson}
\keyword{datasets}
