% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_avatar.R
\docType{data}
\name{avatar_theme_colors}
\alias{avatar_theme_colors}
\title{Avatar Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
avatar_theme_colors
}
\description{
Avatar Theme Color Palette
}
\examples{
avatar_theme_colors

}
\concept{avatar}
\keyword{datasets}
