% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{barbie_box_office}
\alias{barbie_box_office}
\title{Barbie Box Office Data}
\format{
a \code{tibble} with 50 rows and 11 columns
\describe{
\item{\code{date}}{date. Box office date.}
\item{\code{dow}}{factor. Day of the week.}
\item{\code{rank}}{integer. Box office rank.}
\item{\code{daily}}{integer. Daily gross in USD.}
\item{\code{percent_yd}}{numeric. Gross change per day as a percent.}
\item{\code{percent_lw}}{numeric. Gross change per week as a percent.}
\item{\code{theaters}}{integer. Number of theatres.}
\item{\code{avg_per_theater}}{integer. Per theater average gross.}
\item{\code{to_date}}{integer. Gross to date.}
\item{\code{day}}{integer. Day of showing, indexed from 1.}
\item{\code{estimated}}{logical. Flag for if the data is estimated.}
}
}
\source{
Box Office Mojo from \url{https://www.boxofficemojo.com/release/rl1077904129/}
}
\usage{
data("barbie_box_office")
}
\description{
This data contains box office data for the Barbie movie, from its first 50
days (July 21, 2023 to September 9, 2023).
}
\examples{
data(barbie_box_office)
barbie_box_office[, 1:6]
}
\concept{data}
