% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_barbie.R
\docType{data}
\name{barbie_theme_colors}
\alias{barbie_theme_colors}
\title{Barbie Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
barbie_theme_colors
}
\description{
Barbie Theme Color Palette
}
\examples{
barbie_theme_colors

}
\keyword{datasets}
