% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dune.R
\docType{data}
\name{dune_theme_colors}
\alias{dune_theme_colors}
\title{DUNE Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
dune_theme_colors
}
\description{
DUNE Theme Color Palette
}
\examples{
dune_theme_colors

}
\concept{dune}
\keyword{datasets}
