% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_elf.R
\docType{data}
\name{elf_palette}
\alias{elf_palette}
\title{Elf Inspired Color Palette}
\format{
character vector of hex code strings
}
\usage{
elf_palette
}
\description{
Elf Inspired Color Palette
}
\examples{
elf_palette

}
\keyword{datasets}
