% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_elf.R
\docType{data}
\name{elf_theme_colors}
\alias{elf_theme_colors}
\title{Elf Inspired Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
elf_theme_colors
}
\description{
Elf Inspired Theme Color Palette
}
\examples{
elf_theme_colors

}
\concept{elf}
\keyword{datasets}
