% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wes_anderson.R
\docType{data}
\name{french_dispatch_theme_colors}
\alias{french_dispatch_theme_colors}
\title{French Dispatch Inspired Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
french_dispatch_theme_colors
}
\description{
French Dispatch Inspired Theme Color Palette
}
\examples{
french_dispatch_theme_colors

}
\keyword{datasets}
