% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_friends.R
\docType{data}
\name{friends_palette}
\alias{friends_palette}
\title{Friends Color Palette}
\format{
character vector of hex code strings
}
\usage{
friends_palette
}
\description{
Friends Color Palette
}
\examples{
friends_palette

}
\concept{friends}
\keyword{datasets}
