% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_friends.R
\docType{data}
\name{friends_theme_colors}
\alias{friends_theme_colors}
\title{Friends Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
friends_theme_colors
}
\description{
Friends Theme Color Palette
}
\examples{
friends_theme_colors

}
\concept{friends}
\keyword{datasets}
