% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_futurama.R
\docType{data}
\name{futurama_theme_colors}
\alias{futurama_theme_colors}
\title{Futurama Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
futurama_theme_colors
}
\description{
Futurama Theme Color Palette
}
\examples{
futurama_theme_colors

}
\keyword{datasets}
