% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameofthrones.R
\docType{data}
\name{gameofthrones_theme_colors}
\alias{gameofthrones_theme_colors}
\title{Game of Thrones Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
gameofthrones_theme_colors
}
\description{
Game of Thrones Theme Color Palette
}
\examples{
gameofthrones_theme_colors

}
\concept{gameofthrones}
\keyword{datasets}
