% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_godfather.R
\docType{data}
\name{godfather_theme_colors}
\alias{godfather_theme_colors}
\title{Godfather Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
godfather_theme_colors
}
\description{
Godfather Theme Color Palette
}
\examples{
godfather_theme_colors

}
\concept{godfather}
\keyword{datasets}
