% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wes_anderson.R
\docType{data}
\name{grand_budapest_theme_colors}
\alias{grand_budapest_theme_colors}
\title{Grand Budapest Inspired Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
grand_budapest_theme_colors
}
\description{
Grand Budapest Inspired Theme Color Palette
}
\examples{
grand_budapest_theme_colors

}
\concept{wes_anderson}
\keyword{datasets}
