% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hogwarts.R
\docType{data}
\name{gryffindor_theme_colors}
\alias{gryffindor_theme_colors}
\title{Gryffindor Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
gryffindor_theme_colors
}
\description{
Gryffindor Theme Color Palette
}
\examples{
gryffindor_theme_colors

}
\keyword{datasets}
