% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_barbie.R
\name{gt_theme_barbie}
\alias{gt_theme_barbie}
\title{Barbie Inspired Table Theme}
\usage{
gt_theme_barbie(data, ...)
}
\arguments{
\item{data}{a table of class \code{gt_tbl}}

\item{...}{additional arguments to pass to \code{gt::tab_options()}}
}
\value{
a \code{gt_tbl}
}
\description{
Barbie Inspired Table Theme
}
\examples{
gt::exibble |>
  gt::gt() |>
  gt_theme_barbie()
}
