% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hogwarts.R
\docType{data}
\name{hufflepuff_theme_colors}
\alias{hufflepuff_theme_colors}
\title{Hufflepuff Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
hufflepuff_theme_colors
}
\description{
Hufflepuff Theme Color Palette
}
\examples{
hufflepuff_theme_colors

}
\keyword{datasets}
