% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nemo.R
\docType{data}
\name{light_color_nemo}
\alias{light_color_nemo}
\title{#E2F2FD'}
\format{
An object of class \code{character} of length 1.
}
\usage{
light_color_nemo
}
\description{
#E2F2FD'
}
\keyword{datasets}
