% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lordoftherings.R
\docType{data}
\name{lordoftherings_theme_colors}
\alias{lordoftherings_theme_colors}
\title{Lord of the Rings Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
lordoftherings_theme_colors
}
\description{
Lord of the Rings Theme Color Palette
}
\examples{
lordoftherings_theme_colors

}
\keyword{datasets}
