% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nemo.R
\docType{data}
\name{medium_color_nemo}
\alias{medium_color_nemo}
\title{#F6F1F2'}
\format{
An object of class \code{character} of length 1.
}
\usage{
medium_color_nemo
}
\description{
#F6F1F2'
}
\keyword{datasets}
