% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wes_anderson.R
\docType{data}
\name{moonrise_kingdom_theme_colors}
\alias{moonrise_kingdom_theme_colors}
\title{Moonrise Kingdom Inspired Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
moonrise_kingdom_theme_colors
}
\description{
Moonrise Kingdom Inspired Theme Color Palette
}
\examples{
moonrise_kingdom_theme_colors

}
\concept{wes_anderson}
\keyword{datasets}
