% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nemo.R
\docType{data}
\name{nemo_palette}
\alias{nemo_palette}
\title{Finding Nemo Color Palette}
\format{
character vector of hex code strings
}
\usage{
nemo_palette
}
\description{
Finding Nemo Color Palette
}
\examples{
nemo_palette

}
\concept{nemo}
\keyword{datasets}
