% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nemo.R
\docType{data}
\name{nemo_theme_colors}
\alias{nemo_theme_colors}
\title{Finding Nemo Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
nemo_theme_colors
}
\description{
Finding Nemo Theme Color Palette
}
\examples{
nemo_theme_colors

}
\keyword{datasets}
