% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_oppenheimer.R
\docType{data}
\name{oppenheimer_theme_colors}
\alias{oppenheimer_theme_colors}
\title{Oppenheimer Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
oppenheimer_theme_colors
}
\description{
Oppenheimer Theme Color Palette
}
\examples{
oppenheimer_theme_colors

}
\concept{oppenheimer}
\keyword{datasets}
