% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hogwarts.R
\docType{data}
\name{ravenclaw_theme_colors}
\alias{ravenclaw_theme_colors}
\title{Ravenclaw Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
ravenclaw_theme_colors
}
\description{
Ravenclaw Theme Color Palette
}
\examples{
ravenclaw_theme_colors

}
\keyword{datasets}
