% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lordoftherings.R
\docType{data}
\name{sauron_palette}
\alias{sauron_palette}
\title{Sauron Color Palette}
\format{
character vector of hex code strings
}
\usage{
sauron_palette
}
\description{
Sauron Color Palette
}
\examples{
sauron_palette

}
\concept{lordoftherings}
\keyword{datasets}
