% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alien.R
\name{scale_fill_alien}
\alias{scale_fill_alien}
\alias{scale_color_alien}
\title{ALIEN Inspired Color Scales}
\usage{
scale_fill_alien(...)

scale_color_alien(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
ALIEN Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  theme_alien() +
  scale_color_alien() +facet_grid(.~class)

}
