% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dune.R
\name{scale_fill_dune}
\alias{scale_fill_dune}
\alias{scale_color_dune}
\alias{scale_fill_dune_discrete}
\alias{scale_color_dune_discrete}
\title{DUNE Inspired Color Scales}
\usage{
scale_fill_dune(...)

scale_color_dune(...)

scale_fill_dune_discrete(...)

scale_color_dune_discrete(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
DUNE Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  theme_dune() +
  scale_color_dune()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="Mileage Grouped by cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_dune()+
  scale_fill_dune_discrete()

}
\concept{dune}
