% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameofthrones.R
\name{scale_fill_gameofthrones}
\alias{scale_fill_gameofthrones}
\alias{scale_color_gameofthrones}
\title{Game of Thrones Inspired Color Scales}
\usage{
scale_fill_gameofthrones(...)

scale_color_gameofthrones(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
Game of Thrones Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  scale_color_gameofthrones()

}
\concept{gameofthrones}
