% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_godfather.R
\name{scale_fill_godfather}
\alias{scale_fill_godfather}
\alias{scale_color_godfather}
\title{Godfather Inspired Color Scales}
\usage{
scale_fill_godfather(...)

scale_color_godfather(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
Godfather Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  scale_color_godfather() +
  theme_godfather(godfather_font = TRUE)

}
\concept{godfather}
