% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hogwarts.R
\name{scale_hogwarts}
\alias{scale_hogwarts}
\alias{scale_fill_gryffindor}
\alias{scale_color_gryffindor}
\alias{scale_fill_ravenclaw}
\alias{scale_color_ravenclaw}
\alias{scale_fill_hufflepuff}
\alias{scale_color_hufflepuff}
\alias{scale_fill_slytherin}
\alias{scale_color_slytherin}
\title{Harry Potter Inspired Color Scales}
\usage{
scale_fill_gryffindor(...)

scale_color_gryffindor(...)

scale_fill_ravenclaw(...)

scale_color_ravenclaw(...)

scale_fill_hufflepuff(...)

scale_color_hufflepuff(...)

scale_fill_slytherin(...)

scale_color_slytherin(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
Harry Potter Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  scale_color_gryffindor()

}
\concept{hogwarts}
