% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nemo.R
\name{scale_color_nemo_c}
\alias{scale_color_nemo_c}
\alias{scale_fill_nemo_c}
\alias{scale_color_nemo_b}
\alias{scale_fill_nemo_b}
\alias{scale_color_nemo_d}
\alias{scale_fill_nemo_d}
\alias{scale_colour_nemo_d}
\alias{scale_colour_nemo_c}
\alias{scale_colour_nemo_b}
\title{Finding Nemo Inspired Color Scales}
\usage{
scale_color_nemo_c(...)

scale_fill_nemo_c(...)

scale_color_nemo_b(...)

scale_fill_nemo_b(...)

scale_color_nemo_d(...)

scale_fill_nemo_d(...)

scale_colour_nemo_d(...)

scale_colour_nemo_c(...)

scale_colour_nemo_b(...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{ggplot2::binned_scale} for \verb{_b},
\verb{ggplot2::scale_[fill/color]_gradient} for \verb{_c}, or \code{ggplot2::discrete_scale}}
}
\value{
description
}
\description{
Finding Nemo Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  facet_wrap(~(hwy > 29)) +
  theme_nemo(nemo_font = TRUE) +
  scale_fill_nemo_d()

}
