% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_oppenheimer.R
\name{scale_fill_oppenheimer_smoke}
\alias{scale_fill_oppenheimer_smoke}
\alias{scale_color_oppenheimer_smoke}
\alias{scale_fill_oppenheimer_flame}
\alias{scale_color_oppenheimer_flame}
\title{Oppenheimer Inspired Color Scales}
\usage{
scale_fill_oppenheimer_smoke(...)

scale_color_oppenheimer_smoke(...)

scale_fill_oppenheimer_flame(...)

scale_color_oppenheimer_flame(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
Oppenheimer Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  scale_color_oppenheimer_flame()

  #' ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  scale_color_oppenheimer_flame()

}
