% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_snes.R
\name{scale_color_snes_c}
\alias{scale_color_snes_c}
\alias{scale_fill_snes_c}
\alias{scale_color_snes_b}
\alias{scale_fill_snes_b}
\alias{scale_color_snes_d}
\alias{scale_fill_snes_d}
\alias{scale_colour_snes_d}
\alias{scale_colour_snes_c}
\alias{scale_colour_snes_b}
\title{Super Nintendo Inspired Color Scales}
\usage{
scale_color_snes_c(...)

scale_fill_snes_c(...)

scale_color_snes_b(...)

scale_fill_snes_b(...)

scale_color_snes_d(...)

scale_fill_snes_d(...)

scale_colour_snes_d(...)

scale_colour_snes_c(...)

scale_colour_snes_b(...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{ggplot2::binned_scale} for \verb{_b},
\verb{ggplot2::scale_[fill/color]_gradient} for \verb{_c}, or \code{ggplot2::discrete_scale}
for \verb{_d}}
}
\value{
a \code{ggplot} scale object
}
\description{
Super Nintendo Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg, aes(cty)) +
  geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  facet_wrap(~(hwy > 29)) +
  theme_snes() +
  scale_fill_snes_d()

ggplot(mpg) +
  geom_point(aes(x = cty, y  = hwy, color = year)) +
  labs(title = 'Mileage comparisons',
    x = 'City Mileage', y = 'Highway Mileage', color = 'Year') +
  theme_snes() +
  scale_color_snes_c()

}
