% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wes_anderson.R
\name{scale_wes_anderson}
\alias{scale_wes_anderson}
\alias{scale_fill_grand_budapest}
\alias{scale_color_grand_budapest}
\alias{scale_fill_asteroid_city}
\alias{scale_color_asteroid_city}
\alias{scale_fill_french_dispatch}
\alias{scale_color_french_dispatch}
\alias{scale_fill_moonrise_kingdom}
\alias{scale_color_moonrise_kingdom}
\title{Wes Anderson Movies Inspired Color Scales}
\usage{
scale_fill_grand_budapest(...)

scale_color_grand_budapest(...)

scale_fill_asteroid_city(...)

scale_color_asteroid_city(...)

scale_fill_french_dispatch(...)

scale_color_french_dispatch(...)

scale_fill_moonrise_kingdom(...)

scale_color_moonrise_kingdom(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
Wes Anderson Movies Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  scale_color_grand_budapest()

}
\concept{wes_anderson}
