% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zelda.R
\name{scale_fill_zelda}
\alias{scale_fill_zelda}
\alias{scale_color_zelda}
\title{Legend of Zelda Inspired Color Scales}
\usage{
scale_fill_zelda(...)

scale_color_zelda(...)
}
\arguments{
\item{...}{Additional arguments to pass to \verb{ggplot2::scale_[fill/color]_gradient()}}
}
\value{
a \code{ggplot} scale object
}
\description{
Legend of Zelda Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg) +
  geom_point(aes(y = class, x = hwy, color = cyl)) +
  labs(title="MPG by Vehicle Type",
       caption="Source: mpg",
       x = "City Mileage",
       color ="# Cylinders") +
  scale_color_zelda() +
  theme_zelda(zelda_font = TRUE)

}
\concept{zelda}
