% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simpsons.R
\docType{data}
\name{simpsons_theme_colors}
\alias{simpsons_theme_colors}
\title{Simpsons Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
simpsons_theme_colors
}
\description{
Simpsons Theme Color Palette
}
\examples{
simpsons_theme_colors

}
\keyword{datasets}
