% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hogwarts.R
\docType{data}
\name{slytherin_theme_colors}
\alias{slytherin_theme_colors}
\title{Slytherin Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
slytherin_theme_colors
}
\description{
Slytherin Theme Color Palette
}
\examples{
slytherin_theme_colors

}
\concept{hogwarts}
\keyword{datasets}
