% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_snes.R
\docType{data}
\name{snes_palette}
\alias{snes_palette}
\title{Super Nintendo Inspired Color Palette}
\format{
character vector of hex code strings
}
\usage{
snes_palette
}
\description{
Super Nintendo Inspired Color Palette
}
\examples{
snes_palette

}
\concept{snes}
\keyword{datasets}
