% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_snes.R
\docType{data}
\name{snes_theme_colors}
\alias{snes_theme_colors}
\title{Super Nintendo Inspired Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
snes_theme_colors
}
\description{
Super Nintendo Inspired Theme Color Palette
}
\examples{
snes_theme_colors

}
\concept{snes}
\keyword{datasets}
