% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spiderman.R
\docType{data}
\name{spiderman_theme_colors}
\alias{spiderman_theme_colors}
\title{Spiderman Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
spiderman_theme_colors
}
\description{
Spiderman Theme Color Palette
}
\examples{
spiderman_theme_colors

}
\keyword{datasets}
