% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_starwars.R
\docType{data}
\name{starwars_theme_colors}
\alias{starwars_theme_colors}
\title{Star Wars Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
starwars_theme_colors
}
\description{
Star Wars Theme Color Palette
}
\examples{
starwars_theme_colors

}
\keyword{datasets}
