% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_terminator.R
\docType{data}
\name{terminator_theme_colors}
\alias{terminator_theme_colors}
\title{Terminator Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
terminator_theme_colors
}
\description{
Terminator Theme Color Palette
}
\examples{
terminator_theme_colors

}
\keyword{datasets}
