% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alien.R
\name{theme_alien}
\alias{theme_alien}
\title{ALIEN Inspired Theme}
\usage{
theme_alien(alien_font = TRUE, ...)
}
\arguments{
\item{alien_font}{should \code{theme_alien} use Google Font's 'Roboto Condensed'? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
ALIEN Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm',
             color = alien_theme_colors["dark"],
             fill =  alien_theme_colors["light"]) +
 geom_point(color =  alien_theme_colors["medium"]) +
 labs(title = 'ALIEN Scatter Plot') +
 theme_alien()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_alien()

}
