% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_barbie.R
\name{theme_barbie}
\alias{theme_barbie}
\title{Barbie Inspired Theme}
\usage{
theme_barbie(barbie_font = FALSE)
}
\arguments{
\item{barbie_font}{should \code{theme_barbie} use Google Font's Sansita Swashed? Default is \code{FALSE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Barbie Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'Barbie Scatter Plot') +
  theme_barbie(barbie_font = TRUE)

}
