% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dune.R
\name{theme_dune}
\alias{theme_dune}
\title{DUNE Inspired Theme}
\usage{
theme_dune(dune_font = TRUE, ...)
}
\arguments{
\item{dune_font}{should \code{theme_dune} use https://www.dafont.com/dune-rise.font? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
DUNE Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm',
              color = dune_theme_colors["dark"],
              fill =  dune_theme_colors["light"]) +
 geom_point(color =  dune_theme_colors["medium"]) +
 labs(title = 'DUNE Scatter Plot') +
 theme_dune()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="Mileage Grouped by cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_dune()+
  scale_fill_dune_discrete()

}
\concept{dune}
