% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_elf.R
\name{theme_elf}
\alias{theme_elf}
\title{Elf Inspired Theme}
\usage{
theme_elf(elf_font = TRUE, ...)
}
\arguments{
\item{elf_font}{should \code{theme_elf} use custom font? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Elf Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'elf Scatter Plot') +
  theme_elf()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8, color = NA) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_elf() +
  scale_fill_elf_d()


}
