% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_friends.R
\name{theme_friends}
\alias{theme_friends}
\title{Friends tv series Inspired Theme}
\usage{
theme_friends(friends_font = TRUE, ...)
}
\arguments{
\item{friends_font}{should \code{theme_friends} use Google Font's Permanent
Marker? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Friends tv series Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'friends Scatter Plot') +
  theme_friends()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_friends() +
  scale_fill_friends_d()


}
