% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_futurama.R
\name{theme_futurama}
\alias{theme_futurama}
\title{Futurama Inspired Theme}
\usage{
theme_futurama(futurama_font = FALSE, ...)
}
\arguments{
\item{futurama_font}{should \code{theme_futurama} use Google Font's Dongle? Default is \code{FALSE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Futurama Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'Futurama Scatter Plot') +
  theme_futurama(futurama_font = TRUE)

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_futurama(futurama_font = TRUE)

}
