% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameofthrones.R
\name{theme_gameofthrones}
\alias{theme_gameofthrones}
\title{Game of Thrones Inspired Theme}
\usage{
theme_gameofthrones(gameofthrones_font = TRUE, ...)
}
\arguments{
\item{gameofthrones_font}{should \code{theme_gameofthrones} use Google Font's Cinzel? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Game of Thrones Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'Game of Thrones Scatter Plot') +
  theme_gameofthrones()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_gameofthrones()

}
