% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hogwarts.R
\name{theme_gryffindor}
\alias{theme_gryffindor}
\alias{theme_ravenclaw}
\alias{theme_hufflepuff}
\alias{theme_slytherin}
\title{Harry Potter Inspired Themes}
\usage{
theme_gryffindor(gryffindor_font = TRUE, ...)

theme_ravenclaw(ravenclaw_font = TRUE, ...)

theme_hufflepuff(hufflepuff_font = TRUE, ...)

theme_slytherin(slytherin_font = TRUE, ...)
}
\arguments{
\item{gryffindor_font}{should \code{theme_gryffindor} use Google Font's Henny Penny? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}

\item{ravenclaw_font}{should \code{theme_gryffindor} use Google Font's Henny Penny? Default is \code{TRUE}.}

\item{hufflepuff_font}{should \code{theme_gryffindor} use Google Font's Henny Penny? Default is \code{TRUE}.}

\item{slytherin_font}{should \code{theme_gryffindor} use Google Font's Henny Penny? Default is \code{TRUE}.}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Harry Potter Inspired Themes
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm', color = '#890000') +
  geom_point() +
  labs(title = 'Gryffindor Scatter Plot') +
  theme_gryffindor()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_ravenclaw()

}
\concept{hogwarts}
