% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lordoftherings.R
\name{theme_lordoftherings}
\alias{theme_lordoftherings}
\title{Lord of the Rings Inspired Theme}
\usage{
theme_lordoftherings(lordoftherings_font = FALSE)
}
\arguments{
\item{lordoftherings_font}{should \code{theme_lordoftherings} use Google Font's Lugrasimo? Default is \code{FALSE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Lord of the Rings Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'Lord of the Rings Scatter Plot') +
  theme_lordoftherings(lordoftherings_font = TRUE)

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_lordoftherings(lordoftherings_font = TRUE)

}
