% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nemo.R
\name{theme_nemo}
\alias{theme_nemo}
\title{Finding Nemo Inspired Theme}
\usage{
theme_nemo(nemo_font = TRUE, ...)
}
\arguments{
\item{nemo_font}{should \code{theme_nemo} use Google Font's Bowlby One SC? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Finding Nemo Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'Nemo Scatter Plot') +
  theme_nemo()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_nemo() +
  scale_fill_nemo_d()


}
\concept{nemo}
