% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_oppenheimer.R
\name{theme_oppenheimer}
\alias{theme_oppenheimer}
\title{Oppenheimer Inspired Theme}
\usage{
theme_oppenheimer(oppenheimer_font = TRUE, ...)
}
\arguments{
\item{oppenheimer_font}{should \code{theme_oppenheimer} use Google Font's IM Fell English? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Oppenheimer Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'Oppenheimer Scatter Plot') +
  theme_oppenheimer()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_oppenheimer()

}
\concept{oppenheimer}
