% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_starwars.R
\name{theme_starwars}
\alias{theme_starwars}
\title{Star Wars Inspired Theme}
\usage{
theme_starwars(starwars_font = FALSE, ...)
}
\arguments{
\item{starwars_font}{should \code{theme_starwars} use Google Font's IM Fell English? Default is \code{FALSE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Star Wars Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point(color = 'white') +
  labs(title = 'Star Wars Scatter Plot') +
  theme_starwars(starwars_font = TRUE)

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_starwars(starwars_font = TRUE)

}
