% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wes_anderson.R
\name{theme_grand_budapest}
\alias{theme_grand_budapest}
\alias{theme_asteroid_city}
\alias{theme_french_dispatch}
\alias{theme_moonrise_kingdom}
\title{Wes Anderson Movies Inspired Themes}
\usage{
theme_grand_budapest(grand_budapest_font = TRUE, ...)

theme_asteroid_city(asteroid_city_font = TRUE, ...)

theme_french_dispatch(french_dispatch_font = TRUE, ...)

theme_moonrise_kingdom(moonrise_kingdom_font = TRUE, ...)
}
\arguments{
\item{grand_budapest_font}{should \code{theme_grand_budapest} use Google Font's Limelight? Default is \code{TRUE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}

\item{asteroid_city_font}{should \code{theme_asteroid_city} use Google Font's Bungee Shade? Default is \code{TRUE}.}

\item{french_dispatch_font}{should \code{theme_french_dispatch} use Google Font's Montserrat Subrayada? Default is \code{TRUE}.}

\item{moonrise_kingdom_font}{should \code{theme_moonrise_kingdom} use Google Font's Lobster? Default is \code{TRUE}.}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Wes Anderson Movies Inspired Themes
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm', color = '#890000') +
  geom_point() +
  labs(title = 'Grand Budapest Scatter Plot') +
  theme_grand_budapest()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_asteroid_city()

}
