% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zelda.R
\name{theme_zelda}
\alias{theme_zelda}
\title{Legend of Zelda Inspired Theme}
\usage{
theme_zelda(zelda_font = FALSE)
}
\arguments{
\item{zelda_font}{should \code{theme_zelda} use Google Font's IM Fell English? Default is \code{FALSE}.}

\item{...}{additional parameters to pass to \code{ggplot2::theme()}}
}
\value{
a \code{ggplot2} \code{theme} element
}
\description{
Legend of Zelda Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'Zelda Scatter Plot') +
  theme_zelda(zelda_font = TRUE)

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_zelda(zelda_font = TRUE)

}
