% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themepark-package.R
\docType{package}
\name{ThemePark-package}
\alias{ThemePark}
\alias{ThemePark-package}
\title{ThemePark: Themes for 'ggplot2' from Popular Culture}
\description{
Provides 'ggplot2' themes that mirror works from popular culture, such as Barbie, Star Wars, Game of Thrones, and others. The package currently holds 14 themes and a number of corresponding discrete color scales, palettes, and fonts. Each theme (e.g., 'theme_barbie') generates a unique color scheme and font for a 'ggplot2' object that matches the color scheme and font found in the movie, TV show, or video game.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
