% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zelda.R
\docType{data}
\name{zelda_palette}
\alias{zelda_palette}
\title{Legend of Zelda Color Palette}
\format{
character vector of hex code strings
}
\usage{
zelda_palette
}
\description{
Legend of Zelda Color Palette
}
\examples{
zelda_palette

}
\keyword{datasets}
