% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zelda.R
\docType{data}
\name{zelda_theme_colors}
\alias{zelda_theme_colors}
\title{Legend of Zelda Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
zelda_theme_colors
}
\description{
Legend of Zelda Theme Color Palette
}
\examples{
zelda_theme_colors

}
\keyword{datasets}
